requirements.txt
# requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug